requirements.txt
fastapi 
uvicorn 
python-dotenv 
requests
openai==0.28